Address CLI error formatter review
