+++ packages/php-wasm/universal/src/lib/index.ts
@@ -19,6 +19,7 @@ export {
 	prettyPrintFullStackTrace,
 	printResponseDebugDetails,
 } from './error-reporting';
+export { describeError } from './error-formatting';
 export { UnhandledRejectionsTarget } from './wasm-error-reporting';
 export { HttpCookieStore } from './http-cookie-store';
 export type { IteratePhpFilesOptions as IterateFilesOptions } from './iterate-files';
